Rework ParsedTool for reuse outside of the tool shed.
